requirements.txt
streamlit
openai
pandas
python-dotenv
psycopg2-binary